simplify mc2 driver
